Service Selector investigation
