+++ dotnet/src/SemanticKernel.Abstractions/AI/ChatClient/AIServiceChatClient.cs
@@ -1,0 +1,58 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.AI;
+using Microsoft.SemanticKernel.Services;
+
+namespace Microsoft.SemanticKernel.AI.ChatCompletion;
+
+/// <summary>
+/// Allow <see cref="IChatClient"/> to be used as an <see cref="IAIService"/> in a <see cref="IAIServiceSelector"/>
+/// </summary>
+internal class AIServiceChatClient : IAIService, IChatClient
+{
+    private readonly IChatClient _chatClient;
+
+    /// <summary>
+    /// Storage for AI service attributes.
+    /// </summary>
+    internal Dictionary<string, object?> _internalAttributes { get; } = [];
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="AIServiceChatClient"/> class.
+    /// </summary>
+    /// <param name="chatClient">Target <see cref="IChatClient"/>.</param>
+    public AIServiceChatClient(IChatClient chatClient)
+    {
+        Verify.NotNull(chatClient);
+        this._chatClient = chatClient;
+
+        var metadata = this._chatClient.GetService<ChatClientMetadata>();
+        Verify.NotNull(metadata);
+
+        this._internalAttributes[nameof(metadata.ModelId)] = metadata.ModelId;
+        this._internalAttributes[nameof(metadata.ProviderName)] = metadata.ProviderName;
+        this._internalAttributes[nameof(metadata.ProviderUri)] = metadata.ProviderUri;
+    }
+
+    /// <inheritdoc />
+    public IReadOnlyDictionary<string, object?> Attributes => this._internalAttributes;
+
+    /// <inheritdoc />
+    public void Dispose() => this._chatClient.Dispose();
+
+    /// <inheritdoc />
+    public Task<ChatResponse> GetResponseAsync(IList<ChatMessage> chatMessages, ChatOptions? options = null, CancellationToken cancellationToken = default)
+        => this._chatClient.GetResponseAsync(chatMessages, options, cancellationToken);
+
+    /// <inheritdoc />
+    public object? GetService(Type serviceType, object? serviceKey = null)
+        => this._chatClient.GetService(serviceType, serviceKey);
+
+    /// <inheritdoc />
+    public IAsyncEnumerable<ChatResponseUpdate> GetStreamingResponseAsync(IList<ChatMessage> chatMessages, ChatOptions? options = null, CancellationToken cancellationToken = default)
+        => this._chatClient.GetStreamingResponseAsync(chatMessages, options, cancellationToken);
+}
