+++ dotnet/src/SemanticKernel.Abstractions/AI/ChatClient/ChatClientExtensions.cs
@@ -1,0 +1,63 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+using System;
+using System.Diagnostics.CodeAnalysis;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.AI;
+
+namespace Microsoft.SemanticKernel.ChatCompletion;
+
+/// <summary>
+/// Class sponsor that holds extension methods for <see cref="IChatClient"/> interface.
+/// </summary>
+public static class ChatClientExtensions
+{
+    /// <summary>
+    /// Get chat multiple chat message content choices for the prompt and settings.
+    /// </summary>
+    /// <remarks>
+    /// This should be used when the settings request for more than one choice.
+    /// </remarks>
+    /// <param name="chatClient">Target chat client service.</param>
+    /// <param name="prompt">The standardized prompt input.</param>
+    /// <param name="executionSettings">The AI execution settings (optional).</param>
+    /// <param name="kernel">The <see cref="Kernel"/> containing services, plugins, and other state for use throughout the operation.</param>
+    /// <param name="cancellationToken">The <see cref="CancellationToken"/> to monitor for cancellation requests. The default is <see cref="CancellationToken.None"/>.</param>
+    /// <returns>List of different chat message content choices generated by the remote model</returns>
+    public static Task<ChatResponse> GetResponseAsync(
+        this IChatClient chatClient,
+        string prompt,
+        PromptExecutionSettings? executionSettings = null,
+        Kernel? kernel = null,
+        CancellationToken cancellationToken = default)
+    {
+        var chatOptions = executionSettings.ToChatOptions(kernel);
+
+        // Try to parse the text as a chat history
+        if (ChatPromptParser.TryParse(prompt, out var chatHistoryFromPrompt))
+        {
+            var messageList = chatHistoryFromPrompt.ToChatMessageList();
+            return chatClient.GetResponseAsync(messageList, chatOptions, cancellationToken);
+        }
+
+        return chatClient.GetResponseAsync(prompt, chatOptions, cancellationToken);
+    }
+
+    /// <summary>Creates an <see cref="IChatCompletionService"/> for the specified <see cref="IChatClient"/>.</summary>
+    /// <param name="client">The chat client to be represented as a chat completion service.</param>
+    /// <param name="serviceProvider">An optional <see cref="IServiceProvider"/> that can be used to resolve services to use in the instance.</param>
+    /// <returns>
+    /// The <see cref="IChatCompletionService"/>. If <paramref name="client"/> is an <see cref="IChatCompletionService"/>, <paramref name="client"/> will
+    /// be returned. Otherwise, a new <see cref="IChatCompletionService"/> will be created that wraps <paramref name="client"/>.
+    /// </returns>
+    [Experimental("SKEXP0001")]
+    public static IChatCompletionService AsChatCompletionService(this IChatClient client, IServiceProvider? serviceProvider = null)
+    {
+        Verify.NotNull(client);
+
+        return client is IChatCompletionService chatCompletionService ?
+            chatCompletionService :
+            new ChatClientChatCompletionService(client, serviceProvider);
+    }
+}
