+++ dotnet/src/SemanticKernel.Abstractions/AI/ChatCompletion/ChatCompletionServiceExtensions.cs
@@ -128,148 +128,4 @@ public static class ChatCompletionServiceExtensions
             chatClient :
             new ChatCompletionServiceChatClient(service);
     }
-
-    /// <summary>Creates an <see cref="IChatCompletionService"/> for the specified <see cref="IChatClient"/>.</summary>
-    /// <param name="client">The chat client to be represented as a chat completion service.</param>
-    /// <param name="serviceProvider">An optional <see cref="IServiceProvider"/> that can be used to resolve services to use in the instance.</param>
-    /// <returns>
-    /// The <see cref="IChatCompletionService"/>. If <paramref name="client"/> is an <see cref="IChatCompletionService"/>, <paramref name="client"/> will
-    /// be returned. Otherwise, a new <see cref="IChatCompletionService"/> will be created that wraps <paramref name="client"/>.
-    /// </returns>
-    [Experimental("SKEXP0001")]
-    public static IChatCompletionService AsChatCompletionService(this IChatClient client, IServiceProvider? serviceProvider = null)
-    {
-        Verify.NotNull(client);
-
-        return client is IChatCompletionService chatCompletionService ?
-            chatCompletionService :
-            new ChatClientChatCompletionService(client, serviceProvider);
-    }
-
-    /// <summary>Converts a <see cref="ChatMessageContent"/> to a <see cref="ChatMessage"/>.</summary>
-    /// <remarks>This conversion should not be necessary once SK eventually adopts the shared content types.</remarks>
-    internal static ChatMessage ToChatMessage(ChatMessageContent content)
-    {
-        ChatMessage message = new()
-        {
-            AdditionalProperties = content.Metadata is not null ? new(content.Metadata) : null,
-            AuthorName = content.AuthorName,
-            RawRepresentation = content.InnerContent,
-            Role = content.Role.Label is string label ? new ChatRole(label) : ChatRole.User,
-        };
-
-        foreach (var item in content.Items)
-        {
-            AIContent? aiContent = null;
-            switch (item)
-            {
-                case Microsoft.SemanticKernel.TextContent tc:
-                    aiContent = new Microsoft.Extensions.AI.TextContent(tc.Text);
-                    break;
-
-                case Microsoft.SemanticKernel.ImageContent ic:
-                    aiContent =
-                        ic.DataUri is not null ? new Microsoft.Extensions.AI.DataContent(ic.DataUri, ic.MimeType ?? "image/*") :
-                        ic.Uri is not null ? new Microsoft.Extensions.AI.DataContent(ic.Uri, ic.MimeType ?? "image/*") :
-                        null;
-                    break;
-
-                case Microsoft.SemanticKernel.AudioContent ac:
-                    aiContent =
-                        ac.DataUri is not null ? new Microsoft.Extensions.AI.DataContent(ac.DataUri, ac.MimeType ?? "audio/*") :
-                        ac.Uri is not null ? new Microsoft.Extensions.AI.DataContent(ac.Uri, ac.MimeType ?? "audio/*") :
-                        null;
-                    break;
-
-                case Microsoft.SemanticKernel.BinaryContent bc:
-                    aiContent =
-                        bc.DataUri is not null ? new Microsoft.Extensions.AI.DataContent(bc.DataUri, bc.MimeType) :
-                        bc.Uri is not null ? new Microsoft.Extensions.AI.DataContent(bc.Uri, bc.MimeType) :
-                        null;
-                    break;
-
-                case Microsoft.SemanticKernel.FunctionCallContent fcc:
-                    aiContent = new Microsoft.Extensions.AI.FunctionCallContent(fcc.Id ?? string.Empty, fcc.FunctionName, fcc.Arguments);
-                    break;
-
-                case Microsoft.SemanticKernel.FunctionResultContent frc:
-                    aiContent = new Microsoft.Extensions.AI.FunctionResultContent(frc.CallId ?? string.Empty, frc.Result);
-                    break;
-            }
-
-            if (aiContent is not null)
-            {
-                aiContent.RawRepresentation = item.InnerContent;
-                aiContent.AdditionalProperties = item.Metadata is not null ? new(item.Metadata) : null;
-
-                message.Contents.Add(aiContent);
-            }
-        }
-
-        return message;
-    }
-
-    /// <summary>Converts a <see cref="ChatMessage"/> to a <see cref="ChatMessageContent"/>.</summary>
-    /// <remarks>This conversion should not be necessary once SK eventually adopts the shared content types.</remarks>
-    internal static ChatMessageContent ToChatMessageContent(ChatMessage message, Microsoft.Extensions.AI.ChatResponse? response = null)
-    {
-        ChatMessageContent result = new()
-        {
-            ModelId = response?.ModelId,
-            AuthorName = message.AuthorName,
-            InnerContent = response?.RawRepresentation ?? message.RawRepresentation,
-            Metadata = message.AdditionalProperties,
-            Role = new AuthorRole(message.Role.Value),
-        };
-
-        foreach (AIContent content in message.Contents)
-        {
-            KernelContent? resultContent = null;
-            switch (content)
-            {
-                case Microsoft.Extensions.AI.TextContent tc:
-                    resultContent = new Microsoft.SemanticKernel.TextContent(tc.Text);
-                    break;
-
-                case Microsoft.Extensions.AI.DataContent dc when dc.MediaTypeStartsWith("image/"):
-                    resultContent = dc.Data is not null ?
-                        new Microsoft.SemanticKernel.ImageContent(dc.Uri) :
-                        new Microsoft.SemanticKernel.ImageContent(new Uri(dc.Uri));
-                    break;
-
-                case Microsoft.Extensions.AI.DataContent dc when dc.MediaTypeStartsWith("audio/"):
-                    resultContent = dc.Data is not null ?
-                        new Microsoft.SemanticKernel.AudioContent(dc.Uri) :
-                        new Microsoft.SemanticKernel.AudioContent(new Uri(dc.Uri));
-                    break;
-
-                case Microsoft.Extensions.AI.DataContent dc:
-                    resultContent = dc.Data is not null ?
-                        new Microsoft.SemanticKernel.BinaryContent(dc.Uri) :
-                        new Microsoft.SemanticKernel.BinaryContent(new Uri(dc.Uri));
-                    break;
-
-                case Microsoft.Extensions.AI.FunctionCallContent fcc:
-                    resultContent = new Microsoft.SemanticKernel.FunctionCallContent(fcc.Name, null, fcc.CallId, fcc.Arguments is not null ? new(fcc.Arguments) : null);
-                    break;
-
-                case Microsoft.Extensions.AI.FunctionResultContent frc:
-                    resultContent = new Microsoft.SemanticKernel.FunctionResultContent(callId: frc.CallId, result: frc.Result);
-                    break;
-            }
-
-            if (resultContent is not null)
-            {
-                resultContent.Metadata = content.AdditionalProperties;
-                resultContent.InnerContent = content.RawRepresentation;
-                resultContent.ModelId = response?.ModelId;
-                result.Items.Add(resultContent);
-            }
-        }
-
-        return result;
-    }
-
-    internal static List<ChatMessage> ToChatMessageList(ChatHistory chatHistory)
-        => chatHistory.Select(ToChatMessage).ToList();
 }
