+++ dotnet/src/SemanticKernel.Abstractions/AI/ChatCompletion/ChatHistoryExtensions.cs
@@ -5,6 +5,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Extensions.AI;
 
 namespace Microsoft.SemanticKernel.ChatCompletion;
 
@@ -80,4 +81,6 @@ public static class ChatHistoryExtensions
 
         return chatHistory;
     }
+    internal static List<ChatMessage> ToChatMessageList(this ChatHistory chatHistory)
+        => chatHistory.Select(m => m.ToChatMessage()).ToList();
 }
